Moved the rescaling of B1+ maps to the conversion step
