+++ shimmingtoolbox/dicom_to_nifti.py
@@ -3,13 +3,24 @@
 
 from distutils.dir_util import copy_tree
 import json
+import logging
+import nibabel as nib
+import numpy as np
 import os
-import subprocess
 # from dcm2bids.scaffold import scaffold
 import shutil
+import subprocess
 
 from shimmingtoolbox import __dir_config_dcm2bids__
+from shimmingtoolbox.coils.coordinates import get_main_orientation
 from shimmingtoolbox.utils import create_output_dir
+
+logging.basicConfig()
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+GAMMA = 2.675e8  # Proton's gyromagnetic ratio (rad/(T.s))
+SATURATION_FA = 90  # Saturation flip angle hard-coded in TFL B1 mapping sequence (deg)
 
 
 def dicom_to_nifti(path_dicom, path_nifti, subject_id='sub-01', path_config_dcm2bids=__dir_config_dcm2bids__,
@@ -86,6 +97,7 @@ def dicom_to_nifti(path_dicom, path_nifti, subject_id='sub-01', path_config_dcm2
                             digits = '0123456789'
                             fname_new_json = file_parts[0] + 'phasediff' + file_parts[1].lstrip(digits)
                             is_renaming = True
+
             # Rename the json file an nifti file
             if is_renaming:
                 if os.path.exists(os.path.splitext(fname_json)[0] + '.nii.gz'):
@@ -93,6 +105,74 @@ def dicom_to_nifti(path_dicom, path_nifti, subject_id='sub-01', path_config_dcm2
                     fname_nifti_old = os.path.splitext(fname_json)[0] + '.nii.gz'
                     os.rename(fname_nifti_old, fname_nifti_new)
                     os.rename(fname_json, fname_new_json)
+
+    # Go in the RF map folder
+    path_rfmap = os.path.join(path_nifti, 'sub-' + subject_id, 'rfmap')
+    if os.path.exists(path_rfmap):
+        # Make a list of the json files in rfmap folder
+        file_list = []
+        [file_list.append(os.path.join(path_rfmap, f)) for f in os.listdir(path_rfmap) if
+         os.path.splitext(f)[1] == '.json']
+        file_list = sorted(file_list)
+        for path_json_b1 in file_list:
+            # Do nothing if the files have already been processed by Shimming-Toolbox
+            if '_uncombined' in path_json_b1 or '_shimmed' in path_json_b1:
+                continue
+            # Open the json file
+            with open(path_json_b1) as json_file:
+                json_data = json.load(json_file)
+                # Check what B1+ mapping sequence has been used to proceed accordingly
+                # If Siemens' TurboFLASH B1 mapping (dcm2niix cannot separate phase and magnitude for this sequence)
+                if ('SequenceName' in json_data) and 'tfl2d1_16' in json_data['SequenceName']:
+                    path_nii_b1 = path_json_b1.split(".json")[0] + ".nii.gz"
+                    nii_b1 = nib.load(path_nii_b1)
+                    image = scale_tfl_b1(nii_b1.get_fdata(), json_data)
+                    nii_b1.header['datatype'] = 32  # 32 corresponds to complex data
+                    nii_b1.header['aux_file'] = 'Uncombined B1+ maps'
+                    # tfl_rfmap yields bogus affine matrices that need to be fixed to visualize the B1+ maps in FSLeyes
+                    qfac = nii_b1.header['pixdim'][0]
+
+                    if 'ImageOrientationPatientDICOM' not in json_data:
+                        raise KeyError("Missing json tag: 'ImageOrientationPatientDICOM'. Check dcm2niix version.")
+
+                    # These values are inverted in ImageOrientationPatientDICOM. Correcting them to repair affine matrix
+                    json_data['ImageOrientationPatientDICOM'][0] = -json_data['ImageOrientationPatientDICOM'][0]
+                    json_data['ImageOrientationPatientDICOM'][1] = -json_data['ImageOrientationPatientDICOM'][1]
+                    json_data['ImageOrientationPatientDICOM'][5] = -json_data['ImageOrientationPatientDICOM'][5]
+
+                    xa, xb, xc, ya, yb, yc = np.asarray(json_data['ImageOrientationPatientDICOM'])
+
+                    # Compute the rotation matrix from the corrected values
+                    R = [[xa, ya, qfac * (xb * yc - xc * yb)],
+                         [xb, yb, qfac * (xc * ya - xa * yc)],
+                         [xc, yc, qfac * (xa * yb - xb * ya)]]
+
+                    # Build the corrected affine matrix
+                    affine = np.zeros((4, 4))
+                    affine[:3, :3] = R * nii_b1.header['pixdim'][1:4]
+                    affine[3, :] = [0, 0, 0, 1]
+                    affine[:3, 3] = [nii_b1.header['qoffset_x'], nii_b1.header['qoffset_y'], nii_b1.header['qoffset_z']]
+
+                    # Create a new NIfTI object with the corrected affine matrix and the reshuffled/rescaled B1+ maps
+                    nii_b1.header.set_sform(affine)
+                    nii_b1 = nib.Nifti1Image(image, affine, header=nii_b1.header)
+                    nib.save(nii_b1, path_nii_b1)
+
+                    # Save uncombined B1+ maps in a NIfTI file that can now be visualized in FSLeyes
+                    json_data["ImageComments"] = 'Complex uncombined B1+ map (nT/V)'
+                    path_nii_b1_new = path_nii_b1.split('.nii')[0] + '_uncombined.nii' + path_nii_b1.split('.nii')[1]
+                    nib.save(nii_b1, path_nii_b1_new)
+
+                    # Save the associated JSON file
+                    file_json_b1 = open(os.path.join(path_nii_b1_new.split('.nii')[0] + '.json'), mode='w')
+                    json.dump(json_data, file_json_b1)
+
+                    # Remove the old buggy NIfTI and associated JSON files
+                    os.remove(path_nii_b1)
+                    os.remove(path_json_b1)
+                # TODO: Add handling of other B1+ mapping sequences
+
+        logger.info("B1+ NIfTI have been reshuffled and rescaled.")
 
     # if 'win' in sys.platform:
     #     # dcm2bids is broken for windows as a python package so using CLI
@@ -133,3 +213,76 @@ def dicom_to_nifti(path_dicom, path_nifti, subject_id='sub-01', path_config_dcm2
 
     if remove_tmp:
         shutil.rmtree(os.path.join(path_nifti, 'tmp_dcm2bids'))
+
+
+def scale_tfl_b1(image, json_data):
+    """Rescales the magnitude and phase of complex B1 maps acquired with Siemens' standard B1 mapping sequence. Also
+    reorders the data that are shuffled during DICOM to NIfTI conversion.
+    Args:
+        image (numpy.ndarray): Array of dimension (x, y, n_slices, 2*n_channels). First half: magnitude in degree*10.
+        Second half: phase in range [248-3848].
+        json_data (dict): Contains the different fields present in the json file corresponding to the nifti file.
+
+    Returns:
+        numpy.ndarray: Complex rescaled B1 maps (x, y, n_slices, n_channels).
+    """
+    # The number of slices corresponds to the 3rd dimension of the shuffled NIfTI volume.
+    n_slices = image.shape[2]
+    # The number of Tx channels corresponds to the 4th dimension of the shuffled NIfTI of the shuffled NIfTI volume.
+    n_channels = image.shape[3]//2
+
+    # Magnitude values are stored in the first half of the 4th dimension
+    b1_mag = image[:, :, :, :n_channels]
+
+    if b1_mag.min() < 0:
+        raise ValueError("Unexpected negative magnitude values")
+
+    # Phase values are stored in the second half of the 4th dimension. Siemens phase range: [248 - 3848]
+    b1_phase = image[:, :, :, n_channels:]
+
+    # Reorder data shuffled by dm2niix into shape (x, y , n_slices*n_channels)
+    b1_mag_vector = np.zeros((image.shape[0], image.shape[1], n_slices * n_channels))
+    b1_phase_vector = np.zeros((image.shape[0], image.shape[1], n_slices * n_channels))
+    if 'ImageOrientationText' in json_data:
+        orientation = json_data['ImageOrientationText'].upper()
+    else:
+        logger.info("No 'ImageOrientationText' tag. Slice orientation determined from 'ImageOrientationPatientDICOM'.")
+        orientation = get_main_orientation(json_data['ImageOrientationPatientDICOM'])
+
+    # Axial or coronal cases (+ tilted)
+    if orientation[:3] in ['TRA', 'COR']:
+        for i in range(n_channels):
+            b1_mag_vector[:, :, i * n_slices:(i + 1) * n_slices] = b1_mag[:, :, :, i]
+            b1_phase_vector[:, :, i * n_slices:(i + 1) * n_slices] = b1_phase[:, :, :, i]
+    # Sagittal case (+ tilted)
+    elif orientation[:3] == 'SAG':
+        for i in range(n_channels):
+            b1_mag_vector[:, :, i * n_slices:(i + 1) * n_slices] = b1_mag[:, :, ::-1, i]
+            b1_phase_vector[:, :, i * n_slices:(i + 1) * n_slices] = b1_phase[:, :, ::-1, i]
+    else:
+        raise ValueError("Unknown slice orientation")
+
+    # Reorder data shuffled by dm2niix into shape (x, y, n_slices, n_channels)
+    b1_mag_ordered = np.zeros_like(b1_mag)
+    b1_phase_ordered = np.zeros_like(b1_phase)
+
+    for i in range(n_slices):
+        b1_mag_ordered[:, :, i, :] = b1_mag_vector[:, :, i * n_channels:i * n_channels + n_channels]
+        b1_phase_ordered[:, :, i, :] = b1_phase_vector[:, :, i * n_channels:i * n_channels + n_channels]
+
+    # Scale magnitude in nT/V
+    b1_mag_ordered = b1_mag_ordered / 10  # Siemens magnitude values are stored in degrees x10
+    b1_mag_ordered[b1_mag_ordered > 180] = 180  # Values higher than 180 degrees are due to noise
+    # Calculate B1+ efficiency (1ms, pi-pulse) and scale by the ratio of the measured FA to the saturation FA.
+    # Get the Transmission amplifier reference amplitude
+    amplifier_voltage = json_data['TxRefAmp']  # [V]
+    socket_voltage = amplifier_voltage * 10 ** -0.095  # -1.9dB voltage loss from amplifier to coil socket
+    b1_mag_ordered = (b1_mag_ordered / SATURATION_FA) * (np.pi / (GAMMA * socket_voltage * 1e-3)) * 1e9  # nT/V
+
+    # Scale the phase between [-pi, pi]
+    # Remove potential out of range zeros (set them as null phase = 2048)
+    b1_phase_ordered[b1_phase_ordered == 0] = 2048
+    b1_phase_ordered = (b1_phase_ordered - 2048) * np.pi / 1800  # [-pi pi]
+
+    # Return complex B1 maps
+    return b1_mag_ordered * np.exp(1j * b1_phase_ordered)
